INS-2418: rename Hnadler to Dispatcher
