+++ insolar/flow/flow.go
@@ -24,10 +24,10 @@ import (
 
 // Handle is a one-function synchronous process that can call routines to do long processing.
 // IMPORTANT: Asynchronous code is NOT ALLOWED here.
-// To create a new Handle of a given message use handler.NewHandler procedure.
+// To create a new Handle of a given message use dispatcher.NewHandler procedure.
 // After creating a Handle you can register it in MessageBus like this:
 // `h.Bus.MustRegister(insolar.TypeGetObject, createdHandle.WrapBusHandle)`
-// You can find an example in insolar/ladger/artifactmanager/handler.go
+// You can find an example in insolar/ladger/artifactmanager/dispatcher.go
 type Handle func(context.Context, Flow) error
 
 // MakeHandle is a function that constructs new Handle.
@@ -51,7 +51,7 @@ type Procedure interface {
 // Flow will be pasted to all Handles to control execution.
 // This is very important not to blow this interface. Keep it minimal.
 type Flow interface {
-	// Handle gives control to another handle and waits for its return. Consider it "calling" another handler.
+	// Handle gives control to another handle and waits for its return. Consider it "calling" another dispatcher.
 	Handle(context.Context, Handle) error
 
 	// Procedure starts a routine and blocks Handle execution until cancellation happens or routine returns.
